requirements.txt
# Core dependencies
httpx>=0.25.0
pydantic>=2.0.0
eval_type_backport>=0.2.0  # Python 3.9 type annotation support
python-dotenv>=1.0.0
tenacity>=8.2.0
PyYAML>=6.0

# Image processing
Pillow>=10.0.0

# PDF generation
reportlab>=4.0.0

# CLI
typer>=0.9.0
rich>=13.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: async support
anyio>=4.0.0
